feat: implement Dark Mode and Cockpit UI refinement (Feature 012)
- Enforce pure black (#000000) default dashboard background
- Implement HmiTheme and HmiPalette with high-contrast industrial colors
- Add auto-migration logic for legacy layouts
- Implement per-widget font scaling (0.5x - 2.5x)
- Restrict widget palette to ensure legibility against black text

+++ app/src/main/java/com/example/hmi/dashboard/WidgetContainer.kt
@@ -38,7 +38,7 @@ fun WidgetContainer(
     content: @Composable () -> Unit
 ) {
     // FIX: Default to theme's Primary color instead of Surface to match Buttons and ColorPicker preview
-    val baseBg = backgroundColor?.let { Color(it.toInt()) } ?: MaterialTheme.colorScheme.primary
+    val baseBg = backgroundColor?.let { Color(it.toULong()) } ?: MaterialTheme.colorScheme.primary
     val bg = baseBg.copy(alpha = alpha)
     val contentColor = ColorUtils.getContrastColor(baseBg).copy(alpha = alpha)
     val borderColor = contentColor.copy(alpha = 0.3f * alpha)
